员工权限（部门信息）、往来单位（客户兼厂家类别，税率）、采购报价销售（检查重复BOM）、兼容没有cache文件夹、更改菜单名字

+++ application/models/cache_model.php
@@ -117,6 +117,9 @@ class Cache_model extends CI_Model{
 	}
 	
 	public function delsome($key) {
+	    if(!is_dir($this->path)) { //当普通用户登录的时候，没有cache文件
+	        return true;
+        }
 	    if (is_dir($this->path.$key)) {
 		    delete_files($this->path.$key);
 		} else {
